--- a/aten/src/ATen/native/UnfoldBackward.cpp
+++ b/aten/src/ATen/native/UnfoldBackward.cpp
@@ -1,0 +1,27 @@
+#include <ATen/native/UnfoldBackward.h>
+
+namespace at { namespace native {
+
+DEFINE_DISPATCH(unfold_backward_stub);
+
+Tensor unfold_backward(
+  const Tensor& grad,
+  IntArrayRef input_sizes,
+  int64_t dim,
+  int64_t size,
+  int64_t step
+) {
+  auto grad_input = at::zeros(input_sizes, grad.options()).contiguous();
+  auto grad_contiguous = grad.contiguous();
+
+  unfold_backward_stub(
+    grad.device().type(),
+    grad_input,
+    grad_contiguous,
+    dim, size, step
+  );
+
+  return grad_input;
+}
+
+}} // namespace at::native