--- a/aten/src/ATen/native/UnfoldBackward.h
+++ b/aten/src/ATen/native/UnfoldBackward.h
@@ -9,8 +9,8 @@
 namespace at { namespace native {
 
 using unfold_backward_fn = void (*)(
-  Tensor& grad_out,
-  const Tensor& grad_in,
+  Tensor& grad_in,
+  const Tensor& grad,
   int64_t dim,
   int64_t size,
   int64_t step
@@ -19,6 +19,10 @@ using unfold_backward_fn = void (*)(
 DECLARE_DISPATCH(unfold_backward_fn, unfold_backward_stub);
 
 namespace {
+
+// Note on naming: it is unconventional.
+// grad_in does not mean that it is a gradient wrt to input,
+// grad_in/grad_out is just an input/output of unfold_backward kernel.
 
 static TensorIterator _make_unfold_backward_iter_over_grad_out(
   Tensor& grad_out,