… aten/src/ATen/native/native_functions.yaml
… tools/autograd/derivatives.yaml
… torch/testing/_internal/common_methods_invocations.py