`unfold_backward` uses `index_add` which causes regression on CUDA because of the underlying `atomicAdd`, and regression on CPU because of limited parallelization. This PR attempts to replace `index_add` with a custom kernel.

Fixes [#17501](https://github.com/pytorch/pytorch/issues/17501).