@nikitaved 2020-04-17
@albanD , sorry to bother you, but could you please help me with the tests again? ;)
This implementation does not seem to be utterly wrong, and 10/12 autograd tests for unfold get passed. However, the two remaining `test_unfold_lastdim_*` fail. Could you tell me where I could find these? Grep, unfortunately, does not give me any results.