@albanD 2020-04-17
Sure!

These tests are automatically generated in `test_autograd.py` using the `add_test()` function.
The definition of everything that is passed to this function can be found [here](https://github.com/pytorch/pytorch/blob/46288465fe773da6941b2f366a0ad6e59f935d83/torch/testing/_internal/common_methods_invocations.py#L443-L445).

You can rerun these tests by doing `python test_autograd.py TestAutogradDeviceTypeCUDA.test_unfold_lastdim_neg0_cuda` for example. Even though you can't find this name in the file, this name is generated by the add_test() function.