@nikitaved 2020-04-17
Thank you! But objective does the autograd test the backward on?