@albanD 2020-04-17
In the file I link where you specify the test, you also specify the size of the Tensor inputs to the function and a random Tensor of this size is given.