@nikitaved 2020-04-17
Yes, but a need an objective to check the gradient. What is the underlying objective that gives me a scalar, not a tensor?