@albanD 2020-04-17
It computes a full Jacobian of that function.
So it will call one backward for each element in the output.