@nikitaved 2020-04-17
TODO: use `unsqueeze` on `grad_in`.