@nikitaved 2020-04-21
I am playing with a modification which is efficient for cases when `step >= size`.
However, during TensorIterator construction I am getting the following error during autograd testing:
```
RuntimeError: output with shape [1] doesn't match the broadcast shape [91, 3]
```
How come these shapes do not broadcast? The same holds when I use `[1,1]` shape.

The part of the code:
```
  auto grad_out_restrided = grad_out.unsqueeze(-1);
  auto grad_out_strides = ensure_nonempty_vec(grad_out.strides().vec());
  auto grad_out_sizes = ensure_nonempty_vec(grad_out.sizes().vec());
  grad_out_strides[dim] = 0;
  grad_out_sizes[dim] = 1;
  grad_out_restrided = grad_out_restrided.as_strided(grad_out_sizes, grad_out_strides);

  auto iter = TensorIterator();
  iter.dont_compute_common_dtype();
  iter.dont_resize_outputs();
  iter.add_output(grad_out_restrided);
  iter.add_input(grad_in, grad_in.device(), grad_in.scalar_type());
```
`grad_in` has an additional dimension of size `size`, and `grad_in.shape[dim] != grad_out.shape[dim]`, the rest is the same...