@nikitaved 2020-04-24
Up to date benchmarks.
```python
import torch
from IPython import get_ipython

ipython = get_ipython()

#torch.set_num_threads(1)

layer = torch.nn.Conv2d(3, 1, 3)
inp = torch.rand(500, 3, 500, 500)

out = layer(inp)


for c in [3, 5, 7, 9, 11, 21]:
    res = out.unfold(3, c, 1)
    res_sum = res.contiguous().sum()

    print(f"c={c}")
    ipython.magic("timeit res_sum.backward(retain_graph=True)")

    del res

print()

for c in [3, 5, 7, 9, 11, 21]:
    res = out.unfold(3, c, c)
    res_sum = res.contiguous().sum()

    print(f"c={c}")
    ipython.magic("timeit res_sum.backward(retain_graph=True)")

    del res

```
### Master
#### CUDA
```
c=3
105 ms ± 220 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
c=5
119 ms ± 342 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
c=7
133 ms ± 325 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
c=9
147 ms ± 215 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
c=11
161 ms ± 363 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
c=21
231 ms ± 553 µs per loop (mean ± std. dev. of 7 runs, 1 loop each)

c=3
90.1 ms ± 209 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
c=5
93.9 ms ± 134 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
c=7
94 ms ± 240 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
c=9
94 ms ± 154 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
c=11
94 ms ± 130 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
c=21
94.2 ms ± 219 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
```
#### CPU
```
c=3
18.5 s ± 123 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
c=5
28 s ± 507 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
c=7
37.9 s ± 351 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
c=9
47.7 s ± 832 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
c=11
55.7 s ± 85.2 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
c=21
1min 42s ± 1.68 s per loop (mean ± std. dev. of 7 runs, 1 loop each)

c=3
8.64 s ± 75.2 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
c=5
8.67 s ± 70.9 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
c=7
9.4 s ± 436 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
c=9
9.18 s ± 212 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
c=11
9.52 s ± 155 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
c=21
9.46 s ± 325 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)

```

### This PR
#### CUDA
```
c=3                                                                                                                                               
87 ms ± 545 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)                                                                               
c=5                                                                                                                                               
91.6 ms ± 229 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)                                                                             
c=7                                                                                                                                               
96.9 ms ± 253 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)                                                                             
c=9                                                                                                                                               
99.1 ms ± 392 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)                                                                             
c=11                                                                                                                                              
104 ms ± 175 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)                                                                              
c=21                                                                                                                                              
124 ms ± 278 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)                                                                              
                                                                                                                                                  
c=3                                                                                                                                               
85.2 ms ± 183 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)                                                                             
c=5                                                                                                                                               
85.4 ms ± 311 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)                                                                             
c=7                                                                                                                                               
85.8 ms ± 218 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
c=9
85.5 ms ± 222 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
c=11
85.8 ms ± 262 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
c=21
86 ms ± 163 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)

```
#### CPU
```
c=3
4.64 s ± 349 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
c=5
4.38 s ± 155 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
c=7
4.34 s ± 115 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
c=9
4.42 s ± 79.9 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
c=11
4.5 s ± 93.9 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
c=21
4.99 s ± 79.3 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)

c=3
4.1 s ± 168 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
c=5
4.08 s ± 105 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
c=7
4.16 s ± 159 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
c=9
4.22 s ± 164 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
c=11
4.2 s ± 139 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
c=21
4.07 s ± 169 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```